feat: extract logs logic into function and add logs command to cli
