+++ swap/src/bin/asb.rs
@@ -20,10 +20,9 @@ use libp2p::swarm::AddressScore;
 use libp2p::Swarm;
 use swap::common::tracing_util::Format;
 use std::convert::TryInto;
-use std::fs::read_dir;
 use std::net::{IpAddr, Ipv4Addr, SocketAddr};
 use std::sync::Arc;
-use std::{env, io};
+use std::env;
 use structopt::clap;
 use structopt::clap::ErrorKind;
 use swap::asb::command::{parse_args, Arguments, Command};
@@ -31,7 +30,7 @@ use swap::asb::config::{
     initial_setup, query_user_for_initial_config, read_config, Config, ConfigNotInitialized,
 };
 use swap::asb::{cancel, punish, redeem, refund, safely_abort, EventLoop, Finality, KrakenRate};
-use swap::common::{self, check_latest_version};
+use swap::common::{self, check_latest_version, print_or_write_logs};
 use swap::database::{open_db, AccessMode};
 use swap::fs::system_data_dir;
 use swap::network::rendezvous::XmrBtcNamespace;
@@ -40,8 +39,6 @@ use swap::protocol::alice::{run, AliceState};
 use swap::seed::Seed;
 use swap::tor::AuthenticatedClient;
 use swap::{bitcoin, kraken, monero, tor};
-use tokio::fs::{create_dir_all, try_exists, File};
-use tokio::io::{stdout, AsyncBufReadExt, AsyncWriteExt, BufReader, Stdout};
 use tracing_subscriber::filter::LevelFilter;
 
 const DEFAULT_WALLET_NAME: &str = "asb-wallet";
@@ -261,94 +258,7 @@ async fn main() -> Result<()> {
             swap_id,
             redact,
         } => {
-            // use provided directory of default one
-            let default_dir = system_data_dir()?.join("logs");
-            let logs_dir = logs_dir.unwrap_or(default_dir);
-
-            tracing::info!("Reading `*.log` files from `{}`", logs_dir.display());
-
-            // get all files in the directory
-            let log_files = read_dir(&logs_dir)?;
-
-            /// Enum for abstracting over output channels
-            enum OutputChannel {
-                File(File),
-                Stdout(Stdout),
-            }
-
-            /// Conveniance method for writing to either file or stdout
-            async fn write_to_channel(
-                mut channel: &mut OutputChannel,
-                output: &str,
-            ) -> Result<(), io::Error> {
-                match &mut channel {
-                    OutputChannel::File(file) => file.write_all(output.as_bytes()).await,
-                    OutputChannel::Stdout(stdout) => stdout.write_all(output.as_bytes()).await,
-                }
-            }
-
-            // check where we should write to
-            let mut output_channel = match output_path {
-                Some(path) => {
-                    // make sure the directory exists
-                    if !try_exists(&path).await? {
-                        let mut dir_part = path.clone();
-                        dir_part.pop();
-                        create_dir_all(&dir_part).await?;
-                    }
-
-                    tracing::info!("Writing logs to `{}`", path.display());
-
-                    // create/open and truncate file.
-                    // this means we aren't appending which is probably intuitive behaviour
-                    // since we reprint the complete logs anyway
-                    OutputChannel::File(File::create(&path).await?)
-                }
-                None => OutputChannel::Stdout(stdout()),
-            };
-
-            // conveniance method for checking whether we should filter a specific line
-            let filter_by_swap_id: Box<dyn Fn(&str) -> bool> = match swap_id {
-                // if we should filter by swap id, check if the line contains the string
-                Some(swap_id) => {
-                    let swap_id = swap_id.to_string();
-                    Box::new(move |line: &str| line.contains(&swap_id))
-                }
-                // otherwise we let every line pass
-                None => Box::new(|_| true),
-            };
-            
-            // print all lines from every log file. TODO: sort files by date?
-            for entry in log_files {
-                // get the file path
-                let file_path = entry?.path();
-
-                // filter for .log files
-                let file_name = file_path
-                    .file_name()
-                    .and_then(|name| name.to_str())
-                    .unwrap_or("");
-
-                if !file_name.ends_with(".log") {
-                    continue;
-                }
-
-                let buf_reader = BufReader::new(File::open(&file_path).await?);
-                let mut lines = buf_reader.lines();
-
-                // print each line, redacted if the flag is set
-                while let Some(line) = lines.next_line().await? {
-                    // check if we should filter this line
-                    if !filter_by_swap_id(&line) {
-                        continue;
-                    }
-
-                    let line = if redact { common::redact(&line) } else { line };
-                    write_to_channel(&mut output_channel, &line).await?;
-                    // don't forget newlines
-                    write_to_channel(&mut output_channel, "\n").await?;
-                }
-            }
+            print_or_write_logs(logs_dir, output_path, swap_id, redact).await?;
         }
         Command::WithdrawBtc { amount, address } => {
             let bitcoin_wallet = init_bitcoin_wallet(&config, &seed, env_config).await?;
